Refactoring.

+++ src/Modules/SmartThermo.Modules.DataViewer/Models/LimitRelay.cs
@@ -2,12 +2,12 @@
 {
     public class LimitRelay
     {
-        public byte TemperatureThreshold1 { get; set; }
+        public double TemperatureThreshold1 { get; set; }
 
-        public byte HysteresisThreshold1 { get; set; }
+        public double HysteresisThreshold1 { get; set; }
 
-        public byte TemperatureThreshold2 { get; set; }
+        public double TemperatureThreshold2 { get; set; }
 
-        public byte HysteresisThreshold2 { get; set; }
+        public double HysteresisThreshold2 { get; set; }
     }
 }
